+++ CHANGELOG.md
@@ -12,7 +12,13 @@ package updates, you can specify your package dependency using
 
 ## [Unreleased]
 
-*No changes yet.*
+### Additions
+
+- The `firstDelta(against: by:)` and `firstDelta(against:)` methods have been
+  added.  They find where two sequences start to differ.  The
+  `diverges(from: by:)` and `diverges(from:)` methods are the analogs for
+  collections.  The `converges(with: by:)` and `converges(with:)` flip the
+  search to identify where two bidirectional collections start to be the same.
 
 ---